Move MvvmCameraDemo to ExampleBrowser
