+++ Source/Examples/WPF/ExampleBrowser/Examples/MvvmCameraDemo/Observable.cs
@@ -1,0 +1,15 @@
+﻿namespace MvvmCameraDemo
+{
+    using System.ComponentModel;
+    using System.Runtime.CompilerServices;
+
+    public class Observable : INotifyPropertyChanged
+    {
+        public event PropertyChangedEventHandler PropertyChanged;
+
+        protected virtual void OnPropertyChanged([CallerMemberName] string propertyName = null)
+        {
+            this.PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
+        }
+    }
+}
