+++ Source/Examples/WPF/ExampleBrowser/Examples/MvvmCameraDemo/ShellView.xaml.cs
@@ -1,0 +1,19 @@
+﻿// --------------------------------------------------------------------------------------------------------------------
+// <copyright file="ShellView.xaml.cs" company="Helix Toolkit">
+//   Copyright (c) 2014 Helix Toolkit contributors
+// </copyright>
+// --------------------------------------------------------------------------------------------------------------------
+
+namespace MvvmCameraDemo
+{
+    using ExampleBrowser;
+
+    [Example(null,"Synchronized cameras.")]
+    public partial class ShellView
+    {
+        public ShellView()
+        {
+            this.InitializeComponent();
+        }
+    }
+}
