+++ Source/Examples/WPF/ExampleBrowser/Examples/MvvmCameraDemo/ViewportView.xaml.cs
@@ -1,0 +1,22 @@
+﻿// --------------------------------------------------------------------------------------------------------------------
+// <copyright file="ViewportView.xaml.cs" company="Helix Toolkit">
+//   Copyright (c) 2014 Helix Toolkit contributors
+// </copyright>
+// <summary>
+//   Interaction logic for ViewportView.xaml
+// </summary>
+// --------------------------------------------------------------------------------------------------------------------
+
+namespace MvvmCameraDemo
+{
+    /// <summary>
+    /// Interaction logic for ViewportView.xaml
+    /// </summary>
+    public partial class ViewportView
+    {
+        public ViewportView()
+        {
+            this.InitializeComponent();
+        }
+    }
+}
