--- Source/Examples/WPF/MvvmCameraDemo/App.xaml
@@ -1,8 +1,0 @@
-﻿<Application 
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             x:Class="MvvmCameraDemo.App">
-    <Application.Resources>
-
-    </Application.Resources>
-</Application>
