--- Source/Examples/WPF/MvvmCameraDemo/App.xaml.cs
@@ -1,20 +1,0 @@
-﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="App.xaml.cs" company="Helix Toolkit">
-//   Copyright (c) 2014 Helix Toolkit contributors
-// </copyright>
-// --------------------------------------------------------------------------------------------------------------------
-
-namespace MvvmCameraDemo
-{
-    using System.Windows;
-
-    public partial class App : Application
-    {
-        Bootstrapper bootstrapper;
-
-        public App()
-        {
-            bootstrapper = new Bootstrapper();
-        }
-    }
-}
