--- Source/Examples/WPF/MvvmCameraDemo/Bootstrapper.cs
@@ -1,12 +1,0 @@
-﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="Bootstrapper.cs" company="Helix Toolkit">
-//   Copyright (c) 2014 Helix Toolkit contributors
-// </copyright>
-// --------------------------------------------------------------------------------------------------------------------
-
-namespace MvvmCameraDemo
-{
-    public class Bootstrapper : Caliburn.Micro.Bootstrapper<MvvmCameraDemo.ViewModels.ShellViewModel>
-    {
-    }
-}
