--- Source/Examples/WPF/MvvmCameraDemo/ViewModels/ViewportViewModel.cs
@@ -1,36 +1,0 @@
-﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="ViewportViewModel.cs" company="Helix Toolkit">
-//   Copyright (c) 2014 Helix Toolkit contributors
-// </copyright>
-// --------------------------------------------------------------------------------------------------------------------
-
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Windows.Media.Media3D;
-using Caliburn.Micro;
-using System.Diagnostics;
-
-namespace MvvmCameraDemo.ViewModels
-{
-    public class ViewportViewModel : Screen
-    {
-        PerspectiveCamera _camera;
-        public PerspectiveCamera Camera
-        {
-            get { return _camera; }
-            set
-            {
-                _camera = value;
-                NotifyOfPropertyChange(() => this.Camera);
-            }
-        }
-
-
-        public ViewportViewModel()
-        {
-
-        }
-    }
-}
