--- Source/Examples/WPF/MvvmCameraDemo/Views/ShellView.xaml
@@ -1,13 +1,0 @@
-﻿<Window x:Class="MvvmCameraDemo.Views.ShellView"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        Title="MainWindow" Height="350" Width="525">
-    <Grid>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition/>
-            <ColumnDefinition/>
-        </Grid.ColumnDefinitions>
-        <ContentControl Grid.Column="0" x:Name="Viewport1"/>
-        <ContentControl  Grid.Column="1" x:Name="Viewport2"/>
-    </Grid>
-</Window>
