--- Source/Examples/WPF/MvvmCameraDemo/Views/ShellView.xaml.cs
@@ -1,18 +1,0 @@
-﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="ShellView.xaml.cs" company="Helix Toolkit">
-//   Copyright (c) 2014 Helix Toolkit contributors
-// </copyright>
-// --------------------------------------------------------------------------------------------------------------------
-
-namespace MvvmCameraDemo.Views
-{
-    using System.Windows;
-
-    public partial class ShellView : Window
-    {
-        public ShellView()
-        {
-            InitializeComponent();
-        }
-    }
-}
