--- Source/Examples/WPF/MvvmCameraDemo/Views/ViewportView.xaml
@@ -1,34 +1,0 @@
-﻿<UserControl x:Class="MvvmCameraDemo.Views.ViewportView"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:h="clr-namespace:HelixToolkit.Wpf;assembly=HelixToolkit.Wpf"
-             xmlns:local="clr-namespace:MvvmCameraDemo"
-             mc:Ignorable="d" 
-             d:DesignHeight="300" d:DesignWidth="300">
-    <Grid>
-        <Grid.Resources>
-            <local:BindingProxy x:Key="proxy" Data="{Binding}" />
-        </Grid.Resources>
-
-        <Grid.RowDefinitions>
-            <RowDefinition Height="auto"/>
-            <RowDefinition/>
-        </Grid.RowDefinitions>
-        <h:HelixViewport3D Grid.Row="1">
-            
-            <h:HelixViewport3D.Camera>              
-                <PerspectiveCamera 
-                    Position="{Binding Data.Camera.Position,Mode=TwoWay, Source={StaticResource proxy}}"
-                    LookDirection="{Binding Data.Camera.LookDirection,Mode=TwoWay, Source={StaticResource proxy}}"
-                    UpDirection="{Binding Data.Camera.UpDirection,Mode=TwoWay, Source={StaticResource proxy}}"
-                    FieldOfView="{Binding Data.Camera.FieldOfView,Mode=TwoWay, Source={StaticResource proxy}}"
-                    />
-            </h:HelixViewport3D.Camera>
-         
-            <h:CubeVisual3D Fill="Blue"/>
-            <h:DefaultLights/>
-        </h:HelixViewport3D>
-    </Grid>
-</UserControl>
