--- Source/Examples/WPF/MvvmCameraDemo/Views/ViewportView.xaml.cs
@@ -1,36 +1,0 @@
-﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="ViewportView.xaml.cs" company="Helix Toolkit">
-//   Copyright (c) 2014 Helix Toolkit contributors
-// </copyright>
-// <summary>
-//   Interaction logic for ViewportView.xaml
-// </summary>
-// --------------------------------------------------------------------------------------------------------------------
-
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
-
-namespace MvvmCameraDemo.Views
-{
-    /// <summary>
-    /// Interaction logic for ViewportView.xaml
-    /// </summary>
-    public partial class ViewportView : UserControl
-    {
-        public ViewportView()
-        {
-            InitializeComponent();
-        }
-    }
-}
