--- Source/Examples/WPF/MvvmCameraDemo/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Caliburn.Micro" version="1.3.1" targetFramework="net45" />
-</packages>
