+++ Source/HelixToolkit.Wpf.sln
@@ -19,8 +19,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExampleBrowser", "Examples\
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Wpf", "Wpf", "{848BE1AD-560A-4419-BD41-994A1A1CCF1F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MvvmCameraDemo", "Examples\WPF\MvvmCameraDemo\MvvmCameraDemo.csproj", "{090771E2-3B20-4644-94F3-D128D8D41797}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HelixToolkit", "HelixToolkit\HelixToolkit.csproj", "{7C0E987E-CC34-4B19-BA6B-B381AEABB530}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HelixToolkit.Tests", "HelixToolkit.Tests\HelixToolkit.Tests.csproj", "{E86BB7C0-A453-4923-916C-1AA32FF068F9}"
@@ -95,16 +93,6 @@ Global
 		{991A4C2B-D830-41CB-A810-047ADCF69219}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
 		{991A4C2B-D830-41CB-A810-047ADCF69219}.Release|Mixed Platforms.Build.0 = Release|Any CPU
 		{991A4C2B-D830-41CB-A810-047ADCF69219}.Release|x86.ActiveCfg = Release|Any CPU
-		{090771E2-3B20-4644-94F3-D128D8D41797}.Debug|Any CPU.ActiveCfg = Debug|x86
-		{090771E2-3B20-4644-94F3-D128D8D41797}.Debug|Mixed Platforms.ActiveCfg = Debug|x86
-		{090771E2-3B20-4644-94F3-D128D8D41797}.Debug|Mixed Platforms.Build.0 = Debug|x86
-		{090771E2-3B20-4644-94F3-D128D8D41797}.Debug|x86.ActiveCfg = Debug|x86
-		{090771E2-3B20-4644-94F3-D128D8D41797}.Debug|x86.Build.0 = Debug|x86
-		{090771E2-3B20-4644-94F3-D128D8D41797}.Release|Any CPU.ActiveCfg = Release|x86
-		{090771E2-3B20-4644-94F3-D128D8D41797}.Release|Mixed Platforms.ActiveCfg = Release|x86
-		{090771E2-3B20-4644-94F3-D128D8D41797}.Release|Mixed Platforms.Build.0 = Release|x86
-		{090771E2-3B20-4644-94F3-D128D8D41797}.Release|x86.ActiveCfg = Release|x86
-		{090771E2-3B20-4644-94F3-D128D8D41797}.Release|x86.Build.0 = Release|x86
 		{7C0E987E-CC34-4B19-BA6B-B381AEABB530}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{7C0E987E-CC34-4B19-BA6B-B381AEABB530}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7C0E987E-CC34-4B19-BA6B-B381AEABB530}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
@@ -134,6 +122,5 @@ Global
 		{CD249FE2-999C-4AC3-810D-C53ACB2A7DF3} = {848BE1AD-560A-4419-BD41-994A1A1CCF1F}
 		{991A4C2B-D830-41CB-A810-047ADCF69219} = {848BE1AD-560A-4419-BD41-994A1A1CCF1F}
 		{848BE1AD-560A-4419-BD41-994A1A1CCF1F} = {A710CED9-33A6-42F1-8A09-55FEC57B6460}
-		{090771E2-3B20-4644-94F3-D128D8D41797} = {848BE1AD-560A-4419-BD41-994A1A1CCF1F}
 	EndGlobalSection
 EndGlobal
